:truck: Moved dtos out of helpers
